preload extensions in web app
