+++ client/shared/src/api/client/mainthread-api.ts
@@ -1,24 +1,18 @@
 import { Remote, proxy } from 'comlink'
-import { Subscription, from, Observable, of, combineLatest, Subject } from 'rxjs'
-import { fromFetch } from 'rxjs/fetch'
-import { catchError, distinctUntilChanged, map, publishReplay, refCount, switchMap } from 'rxjs/operators'
+import { Subscription, from, Observable, Subject } from 'rxjs'
+import { publishReplay, refCount, switchMap } from 'rxjs/operators'
 import * as sourcegraph from 'sourcegraph'
 
-import { checkOk } from '../../backend/fetch'
 import { registerBuiltinClientCommands } from '../../commands/commands'
-import { ConfiguredExtension, isExtensionEnabled } from '../../extensions/extension'
-import { ExtensionManifest } from '../../extensions/extensionManifest'
-import { areExtensionsSame } from '../../extensions/extensions'
-import { viewerConfiguredExtensions } from '../../extensions/helpers'
 import { PlatformContext } from '../../platform/context'
 import { isSettingsValid } from '../../settings/settings'
-import { asError, isErrorLike } from '../../util/errors'
+import { asError } from '../../util/errors'
 import { FlatExtensionHostAPI, MainThreadAPI } from '../contract'
 import { proxySubscribable } from '../extension/api/common'
 import { NotificationType, PlainNotification } from '../extension/extensionHostApi'
 
 import { ProxySubscription } from './api/common'
-import { observeLanguage, preloadExtensions } from './preload'
+import { getEnabledExtensions } from './enabledExtensions'
 import { updateSettings } from './services/settings'
 
 /** A registered command in the command registry. */
@@ -130,20 +124,6 @@ export const initMainThreadAPI = (
         commandErrors,
     }
 
-    const enabledExtensions = getEnabledExtensions(platformContext)
-
-    // Preload extensions whenever user enabled extensions or the viewed language changes.
-    // Only do this on Sourcegraph web app since the HTTP cache is not shared
-    // between the browser extension's content script (which runs in the code host's context)
-    // and background page.
-    if (platformContext.clientApplication === 'sourcegraph') {
-        subscription.add(
-            combineLatest([enabledExtensions, observeLanguage()]).subscribe(([extensions, languageID]) => {
-                preloadExtensions({ extensions, languages: new Set([languageID]) })
-            })
-        )
-    }
-
     const api: MainThreadAPI = {
         applySettingsEdit: edit => updateSettings(platformContext, edit),
         requestGraphQL: (request, variables) =>
@@ -184,7 +164,7 @@ export const initMainThreadAPI = (
                 return proxySubscribable(from(staticExtensions).pipe(publishReplay(1), refCount()))
             }
 
-            return proxySubscribable(enabledExtensions)
+            return proxySubscribable(getEnabledExtensions(platformContext))
         },
         logEvent: (eventName, eventProperties) => platformContext.telemetryService?.log(eventName, eventProperties),
         logExtensionMessage: (...data) => console.log(...data),
@@ -205,64 +185,4 @@ function defaultShowInputBox(options?: sourcegraph.InputBoxOptions): Promise<str
         const response = prompt(messageFromExtension(options?.prompt ?? ''), options?.value)
         resolve(response ?? undefined)
     })
-}
-
-/**
- * The manifest of an extension sideloaded during local development.
- *
- * Doesn't include {@link ExtensionManifest#url}, as this is added when
- * publishing an extension to the registry.
- * Instead, the bundle URL is computed from the manifest's `main` field.
- */
-interface SideloadedExtensionManifest extends Omit<ExtensionManifest, 'url'> {
-    name: string
-    main: string
-}
-
-export const getConfiguredSideloadedExtension = (baseUrl: string): Observable<ConfiguredExtension> =>
-    fromFetch(`${baseUrl}/package.json`, { selector: response => checkOk(response).json() }).pipe(
-        map(
-            (response: SideloadedExtensionManifest): ConfiguredExtension => ({
-                id: response.name,
-                manifest: {
-                    ...response,
-                    url: `${baseUrl}/${response.main.replace('dist/', '')}`,
-                },
-                rawManifest: null,
-            })
-        )
-    )
-
-function getEnabledExtensions(
-    context: Pick<
-        PlatformContext,
-        'settings' | 'requestGraphQL' | 'sideloadedExtensionURL' | 'getScriptURLForExtension'
-    >
-): Observable<ConfiguredExtension[]> {
-    const sideloadedExtension: Observable<ConfiguredExtension | null> = from(context.sideloadedExtensionURL).pipe(
-        switchMap(url => (url ? getConfiguredSideloadedExtension(url) : of(null))),
-        catchError(error => {
-            console.error('Error sideloading extension', error)
-            return of(null)
-        })
-    )
-
-    return combineLatest([viewerConfiguredExtensions(context), sideloadedExtension, context.settings]).pipe(
-        map(([configuredExtensions, sideloadedExtension, settings]) => {
-            let enabled = configuredExtensions.filter(extension => isExtensionEnabled(settings.final, extension.id))
-            if (sideloadedExtension) {
-                if (!isErrorLike(sideloadedExtension.manifest) && sideloadedExtension.manifest?.publisher) {
-                    // Disable extension with the same ID while this extension is sideloaded
-                    const constructedID = `${sideloadedExtension.manifest.publisher}/${sideloadedExtension.id}`
-                    enabled = enabled.filter(extension => extension.id !== constructedID)
-                }
-
-                enabled.push(sideloadedExtension)
-            }
-            return enabled
-        }),
-        distinctUntilChanged((a, b) => areExtensionsSame(a, b)),
-        publishReplay(1),
-        refCount()
-    )
 }
